[#49]refactor. 캘린더 함수 util함수로 변경
