+++ FE/airbnb/src/components/header/form/FormCheckIn.tsx
@@ -2,10 +2,10 @@ import { useRef, RefObject, useEffect, MouseEvent } from 'react';
 import { useRecoilValue, useResetRecoilState, useSetRecoilState } from 'recoil';
 import styled from 'styled-components';
 import useToggle from '../../../hooks/useToggle';
-import { calendarOpenState, selectDateState } from '../../../recoil/calendarAtom';
-import { isFormOpenedState, selectCheckBoxState } from '../../../recoil/headerAtom';
+import { calendarOpenState, selectDateState } from '../../../recoilStore/calendarAtom';
+import { isFormOpenedState, selectCheckBoxState } from '../../../recoilStore/headerAtom';
 import HoverBlock from '../HoverBlock';
-import { getDateByTime } from './calendar/calendarDateFn';
+import { getDateByTime } from '../../../util/calendarUtils';
 import FormCalendar from './calendar/FormCalendar';
 import FormColumn from './FormColumn';
 import { ReactComponent as DeleteBtn } from '../../../assets/svg/Property 1=x-circle.svg';
@@ -17,13 +17,13 @@ interface Props {
 const FormCheckIn = ({ checkOutRef }: Props) => {
   const checkInRef = useRef<HTMLDivElement>(null);
   const toggleRef = useRef<HTMLDivElement>(null);
+  const selectBox = useRecoilValue(selectCheckBoxState);
   const selectDate = useRecoilValue(selectDateState);
   const resetSelectDate = useResetRecoilState(selectDateState);
-  const selectBox = useRecoilValue(selectCheckBoxState);
   const setIsCalendarOpen = useSetRecoilState(calendarOpenState);
+  const setIsFormOpened = useSetRecoilState(isFormOpenedState);
   const isChekcInSelected = selectBox === 'checkIn';
   const { open } = useToggle({ clickRef: [checkInRef, checkOutRef], toggleRef, isChekcInSelected });
-  const setIsFormOpened = useSetRecoilState(isFormOpenedState);
 
   useEffect(() => {
     setIsCalendarOpen(open);
