+++ FE/airbnb/src/components/header/form/FormLocation.tsx
@@ -2,7 +2,7 @@ import { useRef, useEffect, FormEvent } from 'react';
 import { useRecoilState, useResetRecoilState, useSetRecoilState } from 'recoil';
 import styled from 'styled-components';
 import useToggle from '../../../hooks/useToggle';
-import { isFormOpenedState, locationState } from '../../../recoil/headerAtom';
+import { isFormOpenedState, locationState } from '../../../recoilStore/headerAtom';
 import HoverBlock from '../HoverBlock';
 import FormColumn from './FormColumn';
 import FormLocationToggle from './FormLocationToggle';
